streaming example with core & futures-cpupool

+++ mapreduce/src/main.rs
@@ -5,7 +5,7 @@ use futures::future::Future;
 
 
 fn main() {
-    let data: Vec<u32> = (0..100).collect();
+    let data: Vec<u64> = (0..100).collect();
     let mut ctrl = lib::Delegator::new(data).unwrap();
     let res = ctrl.ship(0).wait();
